Revert branch to the commit directly before the spark branch was created.
This patch was created via:
  git diff --full-index HEAD..944f63fc48725fab3107c6911274290d791f1405

which is exclusive such that the following commit was not reverted:
  commit 944f63fc48725fab3107c6911274290d791f1405
  Author: Ashutosh Chauhan <hashutosh@apache.org>
  Date:   Mon Jul 7 16:29:43 2014 +0000

     HIVE-7079 : Hive logs errors about missing tables when parsing CTE expressions (Navis via Ashutosh Chauhan)

but the commit immediately after was:
  commit 6d1e3dcd1e3ac6d3ccbbd2d90fdfa5159a97c9b3
  Author: Gopal Vijayaraghavan <gopalv@apache.org>
  Date:   Mon Jul 7 21:38:26 2014 +0000

      HIVE-7144 : Use Text writables directly in ORC dictionaries to avoid String allocations (Gopal V, reviewed by Gunther Hagleitner)

+++ ql/src/java/org/apache/hadoop/hive/ql/exec/AutoProgressor.java
@@ -44,7 +44,7 @@ public class AutoProgressor {
   // Name of the class to report for
   String logClassName = null;
   int notificationInterval;
-  long timeout;
+  int timeout;
   Reporter reporter;
 
   class ReporterTask extends TimerTask {
@@ -116,7 +116,7 @@ public class AutoProgressor {
    * @param timeout - when the autoprogressor should stop reporting (in ms)
    */
   AutoProgressor(String logClassName, Reporter reporter,
-      int notificationInterval, long timeout) {
+      int notificationInterval, int timeout) {
     this.logClassName = logClassName;
     this.reporter = reporter;
     this.notificationInterval = notificationInterval;
